Falsk_wytt/requirements.txt
alembic==1.4.2
APScheduler==3.6.3
asn1crypto==1.4.0
asyncio==3.4.3
attrs==20.3.0
autobahn==18.7.1
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.1
constantly==15.1.0
cryptography==3.2.1
cytoolz==0.11.0
dateparser==0.7.0
DateTime==4.3
dnspython==1.16.0
eth-hash==0.2.0
eth-typing==2.2.2
eth-utils==1.9.5
ethereum-abi-utils==0.4.7
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
gevent==20.9.0
geventhttpclient==1.4.4
greenlet==0.4.17
gunicorn==20.0.4
html5lib==1.1
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
numpy==1.18.2
openssl-python==0.1.1
pandas==1.0.3
Pillow==7.0.0
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybithumb==1.0.14
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.9.8
PyHamcrest==2.0.2
pylru==1.2.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pysha3==1.0.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
redis==2.10.6
regex==2020.11.13
requests==2.22.0
rlp==2.0.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.15
supervisor==4.2.1
toolz==0.11.1
Twisted==19.7.0
txaio==20.4.1
tzlocal==2.0.0
urllib3==1.25.8
web3==3.2.0
webencodings==0.5.1
websocket-client==0.57.0
websocket-server==0.4
Werkzeug==1.0.0
zope.event==4.5.0
zope.interface==5.1.2
